requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
umap-learn>=0.5
tqdm>=4.65
jupyter>=1.0
Pillow>=9.5
scipy>=1.10
